+++ src/data/protocols/db/account/load-account-by-email-repository.ts
@@ -1,4 +1,4 @@
-import { AccountModel } from '../../../domain/models/account'
+import { AccountModel } from '../../../../domain/models/account'
 
 export interface LoadAccountByEmailRepository {
   loadByEmail (email: string): Promise<AccountModel | null>
