+++ src/infra/db/mongodb/account/account-mongo-repository.spec.ts
@@ -1,6 +1,6 @@
 import { Collection } from 'mongodb'
 import { mongoHelper } from '../helpers/mongo-helper'
-import { AccountMongoRepository } from './account'
+import { AccountMongoRepository } from './account-mongo-repository'
 
 interface SutInterface {
   sut: AccountMongoRepository
