+++ src/infra/db/mongodb/log/log-mongo-repository.spec.ts
@@ -1,7 +1,7 @@
 import { Collection } from 'mongodb'
-import { LogErrorRepository } from '../../../../data/protocols/db/log-error-repository'
+import { LogErrorRepository } from '../../../../data/protocols/db/log/log-error-repository'
 import { mongoHelper } from '../helpers/mongo-helper'
-import { LogMongoRepository } from './log'
+import { LogMongoRepository } from './log-mongo-repository'
 
 interface SutInterface {
   sut: LogErrorRepository
