+++ src/presentation/controllers/signup/signup-controller.spec.ts
@@ -1,7 +1,7 @@
-import { SignUpController } from './signup'
+import { SignUpController } from './signup-controller'
 import { ServerError } from '../../errors'
 import { AccountModel } from '../../../domain/models/account'
-import { AddAccount } from './signup-protocols'
+import { AddAccount } from './signup-controller-protocols'
 import { HttpRequest } from '../../protocols'
 import { serverError, badRequest, ok } from '../../helpers/http/http-helper'
 import { Validation } from '../../protocols/validation'
